[refactor] 모두의 여행 / 패키지 구조 변경 (#495)
