+++ android/app/src/main/java/com/teamtripdraw/android/ui/model/allTrips/UiTripOfAll.kt
@@ -1,4 +1,4 @@
-package com.teamtripdraw.android.ui.model
+package com.teamtripdraw.android.ui.model.allTrips
 
 data class UiTripOfAll(
     val tripId: Long,
